requirements.txt
alembic==1.7.7
asgiref==3.5.2
BareNecessities==0.2.8
bcrypt==3.2.0
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.0.4
colorama==0.4.4
DBUtils==3.0.2
distlib==0.3.4
Django==4.0.4
django-email-users==1.1.0
django-posts==0.3.0
dnspython==2.2.1
email-validator==1.1.3
fanstatic==1.2
filelock==3.7.1
flake8==4.0.1
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
flask-database==0.2.0
Flask-Ext==0.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
js==1.0
limiter==0.1.2
Mail==2.1.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
npm==0.1.1
optional-django==0.1.0
pdfkit==1.0.0
Pillow==9.1.1
pipenv==2022.6.7
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.5.0
PyMySQL==1.0.2
requests==2.28.1
shutilwhich==1.1.0
six==1.16.0
SQLAlchemy==1.4.34
sqlparse==0.4.2
stripe==4.0.2
token-bucket==0.2.0
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.14.1
virtualenv-clone==0.5.7
WebOb==1.8.7
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.1
